requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1705494584803/work
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.0
alembic==1.13.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
banal==1.0.6
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
defusedxml==0.7.1
dill==0.3.8
docstring_parser==0.16
efficientnet-pytorch==0.7.1
einops==0.7.0
empatches==0.2.3
faiss==1.8.0
filelock==3.9.0
fonttools==4.49.0
FrEIA==0.2
frozenlist==1.4.1
fsspec==2023.5.0
fvcore==0.1.5.post20221221
grad-cam==1.5.2
greenlet==3.0.3
groundingdino==0.1.0
grpcio @ file:///C:/b/abs_35nay_1l8s/croot/grpc-suite_1685746862992/work
huggingface-hub==0.24.6
idna==3.4
imageio==2.34.0
imgaug==0.4.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1703269254275/work
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonargparse==4.28.0
kiwisolver==1.4.5
kornia==0.6.9
lazy_loader==0.3
lightning-utilities==0.11.0
llvmlite==0.42.0
Mako==1.3.2
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1704908347571/work
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.1
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.10.1
protobuf @ file:///D:/bld/protobuf_1681050122068/work/python
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.7
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.28.1
rich==13.7.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
shapely==2.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==1.4.52
supervision==0.19.0rc3
sympy==1.12
tabulate==0.9.0
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1708285739699/work/tensorboard-2.16.2-py3-none-any.whl#sha256=9f2b4e7dad86667615c0e5cd072f1ea8403fc032a299f0072d6f74855775cc45
tensorboard-data-server @ file:///D:/bld/tensorboard-data-server_1695425544157/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=16250db016726523c76d2af77dbfd82881a4c011f2b62e2fe7b508e3e0722450
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.6.12
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.0+cu121
torch-tb-profiler==0.4.3
torchaudio==2.2.0+cu121
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.17.0+cu121
tqdm==4.66.2
transformers==4.44.2
ttach==0.0.3
typeshed_client==2.5.1
typing_extensions==4.8.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==1.26.13
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
wget==3.2
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
